Finished XML: rendered effectively in chrome
Once I auccessfully referenced the .dtd file (and installed an XML extension), there was a persistent error message related to the dtd file @ line 39. I tried updating its name (in both files) loading it in fleet (which is not up to it yet) and so I left it be since it does render. Outside of that, I believe I am in good shape.