Finished dtd file
(Also referenced in Final XML commit) There was a persistent error in VScode that was not responsive to variations on the name of the offending attribute and I was not able to assess the dtd file effectively in fleet. I am not positive that the error is real, the rendering is fine (but I know that browsers don't crumble with every issue).